pyproject.toml
[project]
name = "dsit-litellm-reporting-agent"
version = "0.1.0"
description = "Claude-powered agent for pulling and reporting LiteLLM application usage via REST API"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[project.scripts]
litellm-agent = "agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
